+++ CPSC222_Final_Project_Alexii_Cory_Lizé/CPSC222_StockSim/src/StockGUI/StartScreen.java
@@ -37,49 +37,49 @@ public class StartScreen {
     @FXML
     public void initialize(){
         try {
-            double cA = Double.parseDouble(cycleAmountText.getText());
+            float cA = Float.parseFloat(cycleAmountText.getText());
             cycleAmountSlider.setValue(cA);
         } catch (NumberFormatException e) {
             System.out.println("Invalid number in TextField, using default slider value.");
             cycleAmountSlider.setValue(0);
         }
         try {
-            double cL = Double.parseDouble(cycleLengthText.getText());
+            float cL = Float.parseFloat(cycleLengthText.getText());
             cycleLengthSlider.setValue(cL);
         } catch (NumberFormatException e) {
             System.out.println("Invalid number in TextField, using default slider value.");
             cycleLengthSlider.setValue(0);
         }
         try {
-            double m = Double.parseDouble(moneyText.getText());
+            float m = Float.parseFloat(moneyText.getText());
             moneySlider.setValue(m);
         } catch (NumberFormatException e) {
             System.out.println("Invalid number in TextField, using default slider value.");
             moneySlider.setValue(0);
         }
         try {
-            double p = Double.parseDouble(peopleText.getText());
+            float p = Float.parseFloat(peopleText.getText());
             peopleSlider.setValue(p);
         } catch (NumberFormatException e) {
             System.out.println("Invalid number in TextField, using default slider value.");
             peopleSlider.setValue(0);
         }
         try {
-            double s = Double.parseDouble(stockText.getText());
+            float s = Float.parseFloat(stockText.getText());
             stockSlider.setValue(s);
         } catch (NumberFormatException e) {
             System.out.println("Invalid number in TextField, using default slider value.");
             stockSlider.setValue(0);
         }
         try {
-            double s = Double.parseDouble(graphLengthText.getText());
+            float s = Float.parseFloat(graphLengthText.getText());
             graphLengthSlider.setValue(s);
         } catch (NumberFormatException e) {
             System.out.println("Invalid number in TextField, using default slider value.");
             graphLengthSlider.setValue(0);
         }
         try {
-            double s = Double.parseDouble(eventText.getText());
+            float s = Float.parseFloat(eventText.getText());
             eventSlider.setValue(s);
         } catch (NumberFormatException e) {
             System.out.println("Invalid number in TextField, using default slider value.");
@@ -87,30 +87,30 @@ public class StartScreen {
         }
 
         cycleAmountSlider.valueProperty().addListener((observable, oldValue, newValue) -> {
-            cycleAmountText.setText(String.format("%.0f", newValue.doubleValue()));
+            cycleAmountText.setText(String.format("%.0f", newValue.floatValue()));
         });
         cycleLengthSlider.valueProperty().addListener((observable, oldValue, newValue) -> {
-            cycleLengthText.setText(String.format("%.1f", newValue.doubleValue()));
+            cycleLengthText.setText(String.format("%.1f", newValue.floatValue()));
         });
         moneySlider.valueProperty().addListener((observable, oldValue, newValue) -> {
-            moneyText.setText(String.format("%.0f", newValue.doubleValue()));
+            moneyText.setText(String.format("%.0f", newValue.floatValue()));
         });
         peopleSlider.valueProperty().addListener((observable, oldValue, newValue) -> {
-            peopleText.setText(String.format("%.0f", newValue.doubleValue()));
+            peopleText.setText(String.format("%.0f", newValue.floatValue()));
         });
         stockSlider.valueProperty().addListener((observable, oldValue, newValue) -> {
-            stockText.setText(String.format("%.0f", newValue.doubleValue()));
+            stockText.setText(String.format("%.0f", newValue.floatValue()));
         });
         graphLengthSlider.valueProperty().addListener((observable, oldValue, newValue) -> {
-            graphLengthText.setText(String.format("%.0f", newValue.doubleValue()));
+            graphLengthText.setText(String.format("%.0f", newValue.floatValue()));
         });
         eventSlider.valueProperty().addListener((observable, oldValue, newValue) -> {
-            eventText.setText(String.format("%.0f", newValue.doubleValue()));
+            eventText.setText(String.format("%.0f", newValue.floatValue()));
         });
 
         cycleAmountText.textProperty().addListener((obs, oldVal, newVal) -> {
             try {
-                double value = Double.parseDouble(newVal);
+                float value = Float.parseFloat(newVal);
                 cycleAmountSlider.setValue(value);
             } catch (NumberFormatException e) {
                 // If the input isn't a number (e.g., user types letters), just ignore
@@ -119,7 +119,7 @@ public class StartScreen {
 
         cycleLengthText.textProperty().addListener((obs, oldVal, newVal) -> {
             try {
-                double value = Double.parseDouble(newVal);
+                float value = Float.parseFloat(newVal);
                 cycleLengthSlider.setValue(value);
             } catch (NumberFormatException e) {
                 // If the input isn't a number (e.g., user types letters), just ignore
@@ -128,7 +128,7 @@ public class StartScreen {
 
         moneyText.textProperty().addListener((obs, oldVal, newVal) -> {
             try {
-                double value = Double.parseDouble(newVal);
+                float value = Float.parseFloat(newVal);
                 moneySlider.setValue(value);
             } catch (NumberFormatException e) {
                 // If the input isn't a number (e.g., user types letters), just ignore
@@ -137,7 +137,7 @@ public class StartScreen {
 
         peopleText.textProperty().addListener((obs, oldVal, newVal) -> {
             try {
-                double value = Double.parseDouble(newVal);
+                float value = Float.parseFloat(newVal);
                 peopleSlider.setValue(value);
             } catch (NumberFormatException e) {
                 // If the input isn't a number (e.g., user types letters), just ignore
@@ -146,7 +146,7 @@ public class StartScreen {
 
         stockText.textProperty().addListener((obs, oldVal, newVal) -> {
             try {
-                double value = Double.parseDouble(newVal);
+                float value = Float.parseFloat(newVal);
                 stockSlider.setValue(value);
             } catch (NumberFormatException e) {
                 // If the input isn't a number (e.g., user types letters), just ignore
@@ -154,7 +154,7 @@ public class StartScreen {
         });
         graphLengthText.textProperty().addListener((obs, oldVal, newVal) -> {
             try {
-                double value = Double.parseDouble(newVal);
+                float value = Float.parseFloat(newVal);
                 graphLengthSlider.setValue(value);
             } catch (NumberFormatException e) {
                 // If the input isn't a number (e.g., user types letters), just ignore
@@ -163,7 +163,7 @@ public class StartScreen {
 
         eventText.textProperty().addListener((obs, oldVal, newVal) -> {
             try {
-                double value = Double.parseDouble(newVal);
+                float value = Float.parseFloat(newVal);
                 eventSlider.setValue(value);
             } catch (NumberFormatException e) {
                 // If the input isn't a number (e.g., user types letters), just ignore
@@ -179,9 +179,9 @@ public class StartScreen {
         API.setCycleCount((int) cycleAmountSlider.getValue());
         API.setPeopleAmount((int) peopleSlider.getValue());
         API.setPeopleStartMoney((int) moneySlider.getValue());
-        API.setCycleLength(cycleLengthSlider.getValue());
+        API.setCycleLength((float) cycleLengthSlider.getValue());
         API.setStockStartPrice((int) stockSlider.getValue());
-        API.setEventChance(eventSlider.getValue()/100);
+        API.setEventChance((float) (eventSlider.getValue()/100));
         API.setGraphLength((int) graphLengthSlider.getValue());
 
         FXMLLoader loader = new FXMLLoader(StockGUI.class.getResource("StockGUI.fxml"));
